Source panel should not reappear after sync
This commit moves the GUI/widget logic into update_conversation_view
(it still would be better to put as much GUI logic as we can into
the gui/ directory such that there is a cleaner separation between UI
and other application logic). This makes it such that the source
panel won't re-appear after a sync.

There's an issue, which is that the source panel still is present
until the sync completes for some reason.